+++ CHANGELOG.md
@@ -51,7 +51,7 @@ and this project adheres to
 
 
 <!--
-## [v0.107.3] - 2022-01-19 (APPROX.)
+## [v0.107.3] - 2022-02-08 (APPROX.)
 -->